% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\docType{data}
\name{available_detectors}
\alias{available_detectors}
\title{A list of available detectors}
\format{
An object of class \code{character} of length 5.
}
\usage{
available_detectors
}
\description{
A list of available detectors
}
\keyword{datasets}
