% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\docType{data}
\name{available_layouts}
\alias{available_layouts}
\title{A list of available layouts}
\format{An object of class \code{character} of length 5.}
\usage{
available_layouts
}
\description{
A list of available layouts
}
\keyword{datasets}
