% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{check_detector_avail}
\alias{check_detector_avail}
\title{Checks whether specified detector is available}
\usage{
check_detector_avail(detector_name)
}
\arguments{
\item{detector_name}{The name of the detector}
}
\value{
True or False
}
\description{
Checks whether specified detector is available
}
