% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{check_layout_avail}
\alias{check_layout_avail}
\title{Checks whether specified layout is available}
\usage{
check_layout_avail(layout_name)
}
\arguments{
\item{layout_name}{The name of the layout}
}
\value{
True or False
}
\description{
Checks whether specified layout is available
}
