% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{create_module}
\alias{create_module}
\title{Checks whether detector is available, if so, creates a Detector object}
\usage{
create_module(detector_name)
}
\arguments{
\item{detector_name}{The name of the detector}
}
\value{
Detector object
}
\description{
Checks whether detector is available, if so, creates a Detector object
}
