% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{create_ppp_gaps_col}
\alias{create_ppp_gaps_col}
\title{This is the ppp_gaps_col creation function}
\usage{
create_ppp_gaps_col(detector)
}
\arguments{
\item{detector}{Detector object}
}
\value{
Point pattern dataset
}
\description{
This is the ppp_gaps_col creation function
}
