% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{dead_pix_coords}
\alias{dead_pix_coords}
\title{Extracts a table of dead pixel coordinates from a pixel matrix}
\usage{
dead_pix_coords(pix_matrix)
}
\arguments{
\item{pix_matrix}{pixel matrix with dead pixels flagged with 1}
}
\value{
Table containing dead pixel coordinates
}
\description{
Extracts a table of dead pixel coordinates from a pixel matrix
}
