% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{Dead_Stats}
\alias{Dead_Stats}
\title{Analysis module
A S3 class to represent dead pixels statistics summary}
\usage{
Dead_Stats(
  dead_n = NA,
  module_n = NA,
  module_count_arr = NA,
  module_count = NA,
  avg_dead_mod = NA,
  Chisq_s = NA,
  Chisq_df = NA,
  Chisq_p = NA
)
}
\arguments{
\item{dead_n}{Total number of damaged pixels:}

\item{module_n}{Total number of modules}

\item{module_count_arr}{Count of dead pixels in each quadrat}

\item{module_count}{Count of dead pixels in each quadrat}

\item{avg_dead_mod}{Average number of damaged pixels per module}

\item{Chisq_s}{The Chi-Squared test statistic value}

\item{Chisq_df}{Chi-Squared degrees of freedom}

\item{Chisq_p}{Chi-Squared p-value}
}
\value{
Dead_Stats object
}
\description{
Analysis module
A S3 class to represent dead pixels statistics summary
}
