% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{dead_stats_summary}
\alias{dead_stats_summary}
\title{Count number of damaged pixels overall and in different modules}
\usage{
dead_stats_summary(detector)
}
\arguments{
\item{detector}{Detector object}
}
\value{
A string with damaged pixels overall statitics
}
\description{
Count number of damaged pixels overall and in different modules
}
