% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{Default_Layout}
\alias{Default_Layout}
\title{A S3 class to represent a detector layout.}
\usage{
Default_Layout(name = "Default", detector_width = NA,
  detector_height = NA, module_col_n = NA, module_row_n = NA,
  module_col_sizes = NA, module_row_sizes = NA, gap_col_sizes = NA,
  gap_row_sizes = NA, module_edges_col = NA, module_edges_row = NA,
  detector_inconsistency = NA)
}
\description{
A S3 class to represent a detector layout.
}
\section{Slots}{

\describe{
\item{\code{name}}{detector's name}

\item{\code{detector_width}}{detector's width}

\item{\code{detector_height}}{detector's height}

\item{\code{module_col_n}}{number of columns in the grid of modules}

\item{\code{modulerow.n}}{number of rows in the grid of modules}

\item{\code{module_col_sizes}}{vector with widths of the modules}

\item{\code{module_row_sizes}}{vector with heights of the modules}

\item{\code{gap_col_sizes}}{vector with widths of the gaps}

\item{\code{gap_row_sizes}}{vector with heights of the gaps}

\item{\code{detector_inconsistency}}{counts inconsistencies found in parameters entered}
}}

