% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{derive_layout}
\alias{derive_layout}
\title{Deriving additional layout elements}
\usage{
derive_layout(layout)
}
\arguments{
\item{layout}{Layout object}
}
\value{
Layout object
}
\description{
Conditions additional elements of Layout object that are frequently used later
They are calculated from parameters defined in examples
Matrices that contains xy coordiantes of edges of modules
By definition, edges are part of modules (not part of gaps)
i.e. for each module two pairs: first/last col and first/last row.
}
