% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{detector_consist_check}
\alias{detector_consist_check}
\title{Basic checks if parameters entered (slightly redundant on purpose) add up}
\usage{
detector_consist_check(detector = NA)
}
\arguments{
\item{detector}{Detector object}
}
\description{
Basic checks if parameters entered (slightly redundant on purpose) add up
}
