% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{detector_edges}
\alias{detector_edges}
\title{Defines the coordinates of detector's edges using module and gap sizes
Function is in 1d context to be applied to rows and cols separately.
Edges are inside the modules (first/last row/col of module).}
\usage{
detector_edges(m, g)
}
\arguments{
\item{m}{vector of module sizes}

\item{g}{vectors of gap sizes}
}
\value{
Matrix with the information about the edges
}
\description{
Defines the coordinates of detector's edges using module and gap sizes
Function is in 1d context to be applied to rows and cols separately.
Edges are inside the modules (first/last row/col of module).
}
