% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{detector_summary}
\alias{detector_summary}
\title{Returns a string with the detector summary}
\usage{
detector_summary(detector)
}
\arguments{
\item{detector}{Detector object}
}
\value{
String with the detector summary
}
\description{
Returns a string with the detector summary
}
