% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pixel.R
\name{dist_closest_edge}
\alias{dist_closest_edge}
\title{A function to calclutate closest distance to an edge}
\usage{
dist_closest_edge(x, size)
}
\arguments{
\item{x}{}

\item{size}{something else}
}
\value{
what does this mean?
}
\description{
A function to calclutate closest distance to an edge
}
