% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pixel.R
\name{dist_corner}
\alias{dist_corner}
\title{A function to calculate pixel distances from corners}
\usage{
dist_corner(layout)
}
\arguments{
\item{layout}{Layout object}
}
\value{
Matrix containing parallel maxima from the centre for each pixel
}
\description{
A function to calculate pixel distances from corners
}
