% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pixel.R
\name{dist_edge_col}
\alias{dist_edge_col}
\title{A function to calculate pixel horizontal distance to module edge}
\usage{
dist_edge_col(detector)
}
\arguments{
\item{detector}{Detector object}
}
\value{
distance matrix
}
\description{
A function to calculate pixel horizontal distance to module edge
}
