% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pixel.R
\name{dist_edge_row}
\alias{dist_edge_row}
\title{A function to calculate pixel vertical distance to module edge}
\usage{
dist_edge_row(detector)
}
\arguments{
\item{detector}{Detector object}
}
\value{
distance matrix
}
\description{
A function to calculate pixel vertical distance to module edge
}
