% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{dist_vec}
\alias{dist_vec}
\title{Estimates the distance between vectors v and w}
\usage{
dist_vec(v, w)
}
\arguments{
\item{v}{vector}

\item{w}{vector}
}
\value{
distance between vectors v and w
}
\description{
Estimates the distance between vectors v and w
}
