% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pixel.R
\name{.assign_module}
\alias{.assign_module}
\title{Pixel module
Function assign a module to each dead pixel}
\usage{
.assign_module(detector)
}
\arguments{
\item{detector}{Detector object}
}
\value{
dead_modules
}
\description{
Pixel module
Function assign a module to each dead pixel
}
\keyword{internal}
