% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{.assign_pixel_matrix}
\alias{.assign_pixel_matrix}
\title{Assign dead pixels to a detector}
\usage{
.assign_pixel_matrix(detector, pix_matrix)
}
\arguments{
\item{detector}{Detector object}

\item{pix_matrix}{A pixel matrix}
}
\description{
Assign dead pixels to a detector
}
\keyword{internal}
