% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clump.R
\name{.check_clumps}
\alias{.check_clumps}
\title{Checks if correct clumps were found. If not, finds clumps}
\usage{
.check_clumps(detector, row = NA, col = NA)
}
\arguments{
\item{detector}{Detector object}

\item{row}{Module row number}

\item{col}{Module column number}
}
\value{
detector_events Detector object
}
\description{
Checks if correct clumps were found. If not, finds clumps
}
\keyword{internal}
