% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module.R
\name{.check_select}
\alias{.check_select}
\title{Checks if the selected row and column are within the boundaries of the detector}
\usage{
.check_select(detector, row, col)
}
\arguments{
\item{detector}{Detector object}

\item{row}{module row}

\item{col}{module col}
}
\value{
Boolean
}
\description{
Checks if the selected row and column are within the boundaries of the detector
}
\keyword{internal}
