% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clump.R
\name{.classify_clump}
\alias{.classify_clump}
\title{Clasifies a clump}
\usage{
.classify_clump(layout, x, y)
}
\arguments{
\item{layout}{Layout object}

\item{x}{vector containing the x coordinates of a clump}

\item{y}{vector containing the y coordinates of a clump}
}
\value{
the class of a clump (1 - singleton, 2 - double, 3 - triplet,
  4 - larger cluster, unless it actually has the shape of a line,
  5 (6): vertical line where closest edge is the upper (lower) one,
  7 (8): horizontal line where closest edge is the right (left) one)
}
\description{
Clasifies a clump
}
