% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{.create_ppp_edges_row}
\alias{.create_ppp_edges_row}
\title{This is the create_ppp_edges_row creation function}
\usage{
.create_ppp_edges_row(detector)
}
\arguments{
\item{detector}{Detector object}
}
\value{
Point pattern dataset
}
\description{
This is the create_ppp_edges_row creation function
}
\keyword{internal}
