% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{.derive_detector}
\alias{.derive_detector}
\title{Deriving additional detector elements
Conditions additional elements of Detector object that are frequently used later
They are calculated from parameters defined in examples
Matrices that contains xy coordiantes of edges of modules
By definition, edges are part of modules (not part of gaps)
i.e. for each module two pairs: first/last col and first/last row.}
\usage{
.derive_detector(detector)
}
\arguments{
\item{detector}{Detector object}
}
\value{
Detector object
}
\description{
Deriving additional detector elements
Conditions additional elements of Detector object that are frequently used later
They are calculated from parameters defined in examples
Matrices that contains xy coordiantes of edges of modules
By definition, edges are part of modules (not part of gaps)
i.e. for each module two pairs: first/last col and first/last row.
}
\keyword{internal}
