% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pixel.R
\name{.dist_closest_edge}
\alias{.dist_closest_edge}
\title{A function to calculate closest distance to an edge for a pixel}
\usage{
.dist_closest_edge(x, size)
}
\arguments{
\item{x}{Coordinate of pixel}

\item{size}{Size of module}
}
\value{
distance to closest edge
}
\description{
A function to calculate closest distance to an edge for a pixel
}
