% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module.R
\name{.dist_edge}
\alias{.dist_edge}
\title{Function returns distance of a pixel to module edges.}
\usage{
.dist_edge(xy, module_edges)
}
\arguments{
\item{xy}{Coordinate of pixel}

\item{module_edges}{vector of edges of a module}
}
\value{
tmp Distance to edges
}
\description{
Function returns distance of a pixel to module edges.
}
