% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.extract_detector_parameter}
\alias{.extract_detector_parameter}
\title{Checks whether a detector parameter is in the file string}
\usage{
.extract_detector_parameter(file_string, parameter)
}
\arguments{
\item{file_string}{String of a file context}

\item{parameter}{Detector parameter}
}
\value{
parameter value
}
\description{
Checks whether a detector parameter is in the file string
}
\keyword{internal}
