% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.extract_layout_parameter}
\alias{.extract_layout_parameter}
\title{Checks whether a layout parameter is in the file string}
\usage{
.extract_layout_parameter(file_string, parameter)
}
\arguments{
\item{file_string}{String of a file context}

\item{parameter}{Layout parameter}
}
\value{
parameter value
}
\description{
Checks whether a layout parameter is in the file string
}
