% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{.extract_number}
\alias{.extract_number}
\title{Internal function to convert string values to numbers}
\usage{
.extract_number(s)
}
\arguments{
\item{s}{String expression?}
}
\value{
Numeric value
}
\description{
Internal function to convert string values to numbers
}
