% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{.get_detector_ppps}
\alias{.get_detector_ppps}
\title{Generate detector ppps for edges and gaps}
\usage{
.get_detector_ppps(detector)
}
\arguments{
\item{detector}{Detector object}
}
\value{
a list of ppps for edges and gaps
}
\description{
Generate detector ppps for edges and gaps
}
\keyword{internal}
