% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clump.R
\name{.get_events_mask}
\alias{.get_events_mask}
\title{Generates events mask (a matrix with pixels as 0 and events as 1)}
\usage{
.get_events_mask(detector)
}
\arguments{
\item{detector}{Detector object}
}
\value{
events mask
}
\description{
Generates events mask (a matrix with pixels as 0 and events as 1) indicating if a pixel is in an event
as calculated by \code{find_clumps(detc)}
}
\examples{
detc <- Excalibur_exp_1
detc_with_clumps <- find_clumps(detc)
.get_events_mask(detc_with_clumps)
}
\keyword{internal}
