% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{.get_layout_ppps}
\alias{.get_layout_ppps}
\title{Generate layout ppps}
\usage{
.get_layout_ppps(layout)
}
\arguments{
\item{layout}{Layout object}
}
\value{
a list of ppps for edges and gaps
}
\description{
Generate layout ppps
}
