% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{.get_ppp_dead_module}
\alias{.get_ppp_dead_module}
\title{Generates ppp for the dead pixels for a selected module}
\usage{
.get_ppp_dead_module(layout, row, col)
}
\arguments{
\item{layout}{Layout object}

\item{row}{module row number}

\item{col}{module column number}
}
\value{
ppp of dead pixels
}
\description{
Generates ppp for the dead pixels for a selected module
}
