% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.getmode}
\alias{.getmode}
\title{Returns the mode of a set of data}
\usage{
.getmode(v)
}
\arguments{
\item{v}{set of data}
}
\value{
uniqv the value of the mode
}
\description{
Returns the mode of a set of data
}
\keyword{internal}
