% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clump.R
\name{.mask_to_events}
\alias{.mask_to_events}
\title{Converts mask (dead pixels) to events}
\usage{
.mask_to_events(detector, dead_pix_mask, row = NA, col = NA)
}
\arguments{
\item{detector}{Detector object}

\item{dead_pix_mask}{Dead pixels mask}

\item{row}{Module row number}

\item{col}{Module col number}
}
\value{
list of pixels and events
}
\description{
Converts mask (dead pixels) to events
}
\keyword{internal}
