% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{.matrix_from_xml}
\alias{.matrix_from_xml}
\title{Reads in xml file and returns a pixel matrix}
\usage{
.matrix_from_xml(detector, file_path)
}
\arguments{
\item{detector}{Detector object}

\item{file_path}{Path to the xml file}
}
\value{
Data from an xml file
}
\description{
Reads in xml file and returns a pixel matrix
}
\keyword{internal}
