% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{.norm_vec}
\alias{.norm_vec}
\title{Estimates the norm of a vector}
\usage{
.norm_vec(v)
}
\arguments{
\item{v}{vector}
}
\value{
norm of the vector v
}
\description{
Estimates the norm of a vector
}
\keyword{internal}
