% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{.perform_glm}
\alias{.perform_glm}
\title{A simple wrapper around \code{glm()} with family = binomial(link = logit)}
\usage{
.perform_glm(symb_expr, family = binomial(link = logit))
}
\arguments{
\item{symb_expr}{symbolic description of the linear predictor}

\item{family}{a description of the error distribution}
}
\value{
Fitted model

glm_git fitted model
}
\description{
Calls glm(formula = symb_expr, family = family)
}
\keyword{internal}
