% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{.plot_kfg}
\alias{.plot_kfg}
\title{Plots K, F, G functions}
\usage{
.plot_kfg(ppp_obj, func, file_path = NA, caption = TRUE)
}
\arguments{
\item{ppp_obj}{ppp object}

\item{func}{Function name}

\item{file_path}{Output file path}

\item{caption}{Flag to turn on/off figure caption}
}
\description{
Plots K, F, G functions
}
\keyword{internal}
