% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.tr}
\alias{.tr}
\title{Utils module
Calculates the trace value of a square matrix}
\usage{
.tr(m)
}
\arguments{
\item{m}{A square matrix}
}
\value{
tr The trace value
}
\description{
Utils module
Calculates the trace value of a square matrix
}
\keyword{internal}
