% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clump.R
\name{.xyc_ply_func}
\alias{.xyc_ply_func}
\title{Clasifies clumps with respect to xy coordinates.}
\usage{
.xyc_ply_func(detector, xyc_pixel_df)
}
\arguments{
\item{detector}{Detector object}

\item{xyc_pixel_df}{xyc_pixel_df}
}
\value{
data frame with clasification results
}
\description{
Clasifies clumps with respect to xy coordinates.
}
\keyword{internal}
