% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{Excalibur_Detector}
\alias{Excalibur_Detector}
\title{A S3 class to represent the Excalibur detector.}
\usage{
Excalibur_Detector()
}
\value{
Excalibur detector object
}
\description{
A S3 class to represent the Excalibur detector.
}
\examples{
detc <- Excalibur_Detector()
}
