% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{Excalibur_Layout}
\alias{Excalibur_Layout}
\title{A S3 class to represent the Excalibur detector layout.}
\usage{
Excalibur_Layout()
}
\value{
Excalibur layout object
}
\description{
A S3 class to represent the Excalibur detector layout.
}
