% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pixel.R
\name{get_dead_pix_mask}
\alias{get_dead_pix_mask}
\title{Creates a mask matrix of dead pixels}
\usage{
get_dead_pix_mask(detector)
}
\arguments{
\item{detector}{Detector object}
}
\value{
dead pixel mask
}
\description{
Converts the pix_dead attribute of a detector (NX2 list) to a matrix of 1 and 0 (1 for dead pixel)
}
\examples{
# Create a detector
detector_pilatus <- create_detector("Pilatus")
# Load a pixel matrix
file_path <-  system.file("extdata", "Pilatus", "badpixel_mask.tif",
                         package ="detectorchecker")
detector_pilatus <- load_pix_matrix(
 detector = detector_pilatus, file_path = file_path)
# Calculate dead pixel mask
get_dead_pix_mask(detector_pilatus)
}
