% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{get_dead_stats}
\alias{get_dead_stats}
\title{Generate summary of damaged pixels}
\usage{
get_dead_stats(detector)
}
\arguments{
\item{detector}{Detector object}
}
\value{
Dead_Stats object
}
\description{
Generate summary of damaged pixels
}
