% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clump.R
\name{get_events_matrix}
\alias{get_events_matrix}
\title{Generates events matrix for selected events}
\usage{
get_events_matrix(detector, incl_event_list = NA)
}
\arguments{
\item{detector}{Detector object}

\item{incl_event_list}{a list of events to be included}
}
\value{
Events matrix
}
\description{
Generates events matrix for selected events
}
