% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clump.R
\name{glm_events_ctr_linf}
\alias{glm_events_ctr_linf}
\title{Fits events parallel maxima from the centre using glm}
\usage{
glm_events_ctr_linf(detector, incl_event_list = NA)
}
\arguments{
\item{detector}{Detector object}

\item{incl_event_list}{a list of events to be included}
}
\value{
Fitted model
}
\description{
Fits events parallel maxima from the centre using glm
}
