% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clump.R
\name{glm_events_dist_corner}
\alias{glm_events_dist_corner}
\title{Fits events distances to the nearest corner using glm}
\usage{
glm_events_dist_corner(detector, incl_event_list = NA)
}
\arguments{
\item{detector}{Detector object}

\item{incl_event_list}{a list of events to be included}
}
\value{
Fitted model
}
\description{
Fits events distances to the nearest corner using glm
}
