% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clump.R
\name{glm_events_dist_edge_col}
\alias{glm_events_dist_edge_col}
\title{Fits events distances from the module edges by column using glm}
\usage{
glm_events_dist_edge_col(detector, incl_event_list = NA)
}
\arguments{
\item{detector}{Detector object}

\item{incl_event_list}{a list of events to be included}
}
\value{
Fitted model
}
\description{
Fits events distances from the module edges by column using glm
}
