% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{glm_pixel_ctr_eucl}
\alias{glm_pixel_ctr_eucl}
\title{Predict dead pixels from the pixel's euclidean distance from the detector center}
\usage{
glm_pixel_ctr_eucl(detector)
}
\arguments{
\item{detector}{Detector object}
}
\value{
Fitted model
}
\description{
Fit a logistic regression model using \code{glm}.
Predicts dead pixels from the pixel's euclidean distance from the detector center
}
\examples{
glm_pixel_ctr_eucl(PerkinElmerFull_exp_1)
}
