% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{glm_pixel_ctr_linf}
\alias{glm_pixel_ctr_linf}
\title{Predict dead pixels from the pixel's parallel maxima}
\usage{
glm_pixel_ctr_linf(detector)
}
\arguments{
\item{detector}{Detector object}
}
\value{
Fitted model
}
\description{
Fit a logistic regression model using \code{glm}.
Predicts dead pixels from the pixel's parallel maxima
}
\examples{
# Create a detector
detector_pilatus <- create_detector("Pilatus")
# Load a pixel matrix
file_path <-  system.file("extdata", "Pilatus", "badpixel_mask.tif",
                         package ="detectorchecker")
detector_pilatus <- load_pix_matrix(
 detector = detector_pilatus, file_path = file_path)
# Fit logistic regression model
glm_pixel_ctr_linf(detector_pilatus)
}
