% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{glm_pixel_dist_corner}
\alias{glm_pixel_dist_corner}
\title{Predict dead pixels from pixel distances to the nearest corner}
\usage{
glm_pixel_dist_corner(detector)
}
\arguments{
\item{detector}{Detector object}
}
\value{
Fitted model
}
\description{
Fit a logistic regression model using \code{glm}.
Predict dead pixels from pixel distances to the nearest corner
}
\examples{
glm_pixel_dist_corner(PerkinElmerFull_exp_1)
}
