% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{glm_pixel_dist_edge_min}
\alias{glm_pixel_dist_edge_min}
\title{Fits pixel distances to the nearest sub-panel edge using glm}
\usage{
glm_pixel_dist_edge_min(detector)
}
\arguments{
\item{detector}{Detector object}
}
\value{
Fitted model
}
\description{
Fits pixel distances to the nearest sub-panel edge using glm
}
