% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{glm_pixel_dist_edge_row}
\alias{glm_pixel_dist_edge_row}
\title{Fits pixel distances from module edges by row using glm}
\usage{
glm_pixel_dist_edge_row(detector)
}
\arguments{
\item{detector}{Detector object}
}
\value{
Fitted model
}
\description{
Fits pixel distances from module edges by row using glm
}
