% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pixel.R
\name{inconsist_dead_detector}
\alias{inconsist_dead_detector}
\title{Counts damaged pixel locations (dead_data) outside detector (detector)
  and in gaps between modules and give warnings}
\usage{
inconsist_dead_detector(dead_data, detector)
}
\arguments{
\item{dead_data}{Dead pixel locations}

\item{detector}{Detector object}
}
\value{
Inconsistency message
}
\description{
Counts damaged pixel locations (dead_data) outside detector (detector)
  and in gaps between modules and give warnings
}
