% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pixel.R
\name{inconsist_dead_layout}
\alias{inconsist_dead_layout}
\title{Counts damaged pixel locations (dead_data) outside detector (layout)
  and in gaps between modules and give warnings}
\usage{
inconsist_dead_layout(dead_data, layout)
}
\arguments{
\item{dead_data}{Dead pixel locations}

\item{layout}{Layout object}
}
\value{
Inconsistency message
}
\description{
Counts damaged pixel locations (dead_data) outside detector (layout)
  and in gaps between modules and give warnings
}
