% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{is.detector}
\alias{is.detector}
\title{Check that \code{x} is an S3 detector class}
\usage{
is.detector(x)
}
\arguments{
\item{x}{Any variable}
}
\value{
True if x is an instance of detector
}
\description{
Check that \code{x} is an S3 detector class
}
\examples{
x <- Excalibur_Detector()
is.detector(x)
}
