% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{Layout.consistancyCheck}
\alias{Layout.consistancyCheck}
\title{Layout consistency checks}
\usage{
Layout.consistancyCheck(layout = NA)
}
\arguments{
\item{layout}{Layout object}
}
\description{
Basic checks if parameters entered (slightly redundant on purpose) add up
}
