% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{Layout.Default}
\alias{Layout.Default}
\title{A S3 class to represent a detector layout.}
\usage{
Layout.Default(name = "Default", detector.width = NA,
  detector.height = NA, module.col.n = NA, module.row.n = NA,
  module.col.sizes = NA, module.row.sizes = NA, gap.col.sizes = NA,
  gap.row.sizes = NA, module.edges.col = NA, module.edges.row = NA,
  detector.inconsistency = NA)
}
\description{
A S3 class to represent a detector layout.
}
\section{Slots}{

\describe{
\item{\code{name}}{detector's name}

\item{\code{detector.width}}{detector's width}

\item{\code{detector.height}}{detector's height}

\item{\code{module.col.n}}{number of columns in the grid of modules}

\item{\code{module.row.n}}{number of rows in the grid of modules}

\item{\code{module.col.sizes}}{vector with widths of the modules}

\item{\code{module.row.sizes}}{vector with heights of the modules}

\item{\code{gap.col.sizes}}{vector with widths of the gaps}

\item{\code{gap.row.sizes}}{vector with heights of the gaps}

\item{\code{detector.inconsistency}}{counts inconsistencies found in parameters entered}
}}

