% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{Layout.PerkinElmerFull}
\alias{Layout.PerkinElmerFull}
\title{A S3 class to represent the PerkinElmerFull detector layout.}
\usage{
Layout.PerkinElmerFull()
}
\description{
A S3 class to represent the PerkinElmerFull detector layout.
}
