% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{layout_consist_check}
\alias{layout_consist_check}
\title{Basic checks if parameters entered (slightly redundant on purpose) add up}
\usage{
layout_consist_check(layout = NA)
}
\arguments{
\item{layout}{Layout object}
}
\description{
Basic checks if parameters entered (slightly redundant on purpose) add up
}
