% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{layout_summary}
\alias{layout_summary}
\title{Returns a string with the layout summary}
\usage{
layout_summary(layout)
}
\arguments{
\item{layout}{Layout object}
}
\value{
String with the layout summary
}
\description{
Returns a string with the layout summary
}
