% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{matrix_from_hdf}
\alias{matrix_from_hdf}
\title{Reads in hdf file(s) and returns a pixel matrix}
\usage{
matrix_from_hdf(layout, file_path)
}
\arguments{
\item{layout}{Layout object}

\item{file_path}{A list of paths to hdf files. Must be in the correct order.}
}
\value{
Data of a combined dataset from hdf files
}
\description{
Reads in hdf file(s) and returns a pixel matrix
}
