% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{matrix_from_tiff}
\alias{matrix_from_tiff}
\title{I/O module
Reads in tiff file and returns a pixel matrix}
\usage{
matrix_from_tiff(layout, file_path)
}
\arguments{
\item{layout}{Layout object}

\item{file_path}{Path to the tiff file}
}
\value{
Pixel matrix with dead pixels flagged with 1
}
\description{
I/O module
Reads in tiff file and returns a pixel matrix
}
