% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{orient_dist_vec}
\alias{orient_dist_vec}
\title{Calculates distance and orientation of the oriented vector between two points
in order of the second pointing to first (reflecting nearest neighbour (nn) framework)
v, w point coordinates indicating vectors wrt to the origin.
Values: distance and orientation (in [0,360) degrees) of w pointing towards v.}
\usage{
orient_dist_vec(v, w)
}
\arguments{
\item{v}{vector}

\item{w}{vector}
}
\value{
distance and orientation of the oriented vector between two points
}
\description{
Calculates distance and orientation of the oriented vector between two points
in order of the second pointing to first (reflecting nearest neighbour (nn) framework)
v, w point coordinates indicating vectors wrt to the origin.
Values: distance and orientation (in [0,360) degrees) of w pointing towards v.
}
