% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{orientcolfct}
\alias{orientcolfct}
\title{Calculates orientation of the oriented vector between two points
in order of the second pointing to first (reflecting nearest neighbour (nn) framework)}
\usage{
orientcolfct(b)
}
\arguments{
\item{b}{vector with elements 1:2 correspoding to first point and 3:4 corresponding to the second}
}
\value{
orientation of the oriented vector between two points
}
\description{
Calculates orientation of the oriented vector between two points
in order of the second pointing to first (reflecting nearest neighbour (nn) framework)
}
