% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{PerkinElmerCropped1600_Detector}
\alias{PerkinElmerCropped1600_Detector}
\title{A S3 class to represent the PerkinElmerCropped1600 detector.}
\usage{
PerkinElmerCropped1600_Detector()
}
\value{
PerkinElmerCropped1600 detector object
}
\description{
A S3 class to represent the PerkinElmerCropped1600 detector.
}
