% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{PerkinElmerFull_Detector}
\alias{PerkinElmerFull_Detector}
\title{A S3 class to represent the PerkinElmerFull detector.}
\usage{
PerkinElmerFull_Detector()
}
\value{
PerkinElmerFul detector object
}
\description{
A S3 class to represent the PerkinElmerFull detector.
}
