% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{PerkinElmerFull_Layout}
\alias{PerkinElmerFull_Layout}
\title{A S3 class to represent the PerkinElmerFull detector layout.}
\usage{
PerkinElmerFull_Layout()
}
\value{
PerkinElmerFul layout object
}
\description{
A S3 class to represent the PerkinElmerFull detector layout.
}
