% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{PerkinElmerRefurbished_Detector}
\alias{PerkinElmerRefurbished_Detector}
\title{A S3 class to represent the PerkinElmerRefurbished detector.}
\usage{
PerkinElmerRefurbished_Detector()
}
\value{
PerkinElmerRefurbished detector object
}
\description{
A S3 class to represent the PerkinElmerRefurbished detector.
}
