% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{PerkinElmerRefurbished_Layout}
\alias{PerkinElmerRefurbished_Layout}
\title{A S3 class to represent the PerkinElmerRefurbished detector layout.}
\usage{
PerkinElmerRefurbished_Layout()
}
\value{
PerkinElmerRefurbished layout object
}
\description{
A S3 class to represent the PerkinElmerRefurbished detector layout.
}
