% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{Pilatus_Detector}
\alias{Pilatus_Detector}
\title{A S3 class to represent the PerkinElmerRefurbished detector.}
\usage{
Pilatus_Detector()
}
\value{
Pilatus detector object
}
\description{
A S3 class to represent the PerkinElmerRefurbished detector.
}
