% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{Pilatus_Layout}
\alias{Pilatus_Layout}
\title{A S3 class to represent the PerkinElmerRefurbished detector layout.}
\usage{
Pilatus_Layout()
}
\value{
Pilatus layout object
}
\description{
A S3 class to represent the PerkinElmerRefurbished detector layout.
}
