% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pixel.R
\name{pixel_dist_ctr_eucl}
\alias{pixel_dist_ctr_eucl}
\title{A function to calculate euclidean distance from the centre}
\usage{
pixel_dist_ctr_eucl(layout)
}
\arguments{
\item{layout}{Layout object}
}
\value{
Matrix containing euclidean distances from the centre for each
  pixel
}
\description{
A function to calculate euclidean distance from the centre
}
