% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pixel.R
\name{pixel_dist_ctr_linf}
\alias{pixel_dist_ctr_linf}
\title{A function to calculate parallel maxima from the centre for each pixel}
\usage{
pixel_dist_ctr_linf(detector)
}
\arguments{
\item{detector}{Detector object}
}
\value{
Matrix of parallel maxima
}
\description{
A function to calculate parallel maxima from the centre for each pixel
}
