% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_arrows}
\alias{plot_arrows}
\title{A function to plot NN oriented arrrows}
\usage{
plot_arrows(ppp_obj, caption, file_path = NA)
}
\arguments{
\item{ppp_obj}{ppp object}

\item{caption}{caption of the figure}

\item{file_path}{file path}
}
\description{
A function to plot NN oriented arrrows
}
