% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_counts}
\alias{plot_counts}
\title{A function to plot NN oriented arrrows}
\usage{
plot_counts(module_count_arr, caption, file_path = NA)
}
\arguments{
\item{module_count_arr}{Counts per array}

\item{caption}{caption of the figure}

\item{file_path}{file path}
}
\description{
A function to plot NN oriented arrrows
}
