% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_density}
\alias{plot_density}
\title{Plots module
Plots ppp object's density}
\usage{
plot_density(ppp_obj, caption, file_path = NA, adjust = 0.25)
}
\arguments{
\item{ppp_obj}{ppp object}

\item{caption}{caption of the figure}

\item{file_path}{file path}

\item{adjust}{Kernel density bandwidth}
}
\description{
Plots module
Plots ppp object's density
}
