% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{plot_detector}
\alias{plot_detector}
\title{Plot detector}
\usage{
plot_detector(detector, file_path = NA, caption = TRUE)
}
\arguments{
\item{detector}{Detector object}

\item{file_path}{Output file path}

\item{caption}{Flag to turn on/off figure caption}
}
\description{
Plot detector
}
