% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{plot_detector_damaged}
\alias{plot_detector_damaged}
\title{A function to plot detector with damaged pixels}
\usage{
plot_detector_damaged(detector, file_path = NA, caption = TRUE)
}
\arguments{
\item{detector}{Detector object}

\item{file_path}{Output file path}

\item{caption}{Flag to turn on/off figure caption}
}
\description{
A function to plot detector with damaged pixels
}
