% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{plot_detector_density}
\alias{plot_detector_density}
\title{A function to plot densities of dead pixels of detector or module}
\usage{
plot_detector_density(
  detector,
  file_path = NA,
  adjust = 1,
  row = NA,
  col = NA,
  caption = TRUE,
  color = topo.colors(50)
)
}
\arguments{
\item{detector}{Detector object}

\item{file_path}{Output file path}

\item{adjust}{Kernel density bandwidth}

\item{row}{Module row number}

\item{col}{Module column number}

\item{caption}{Flag to turn on/off figure caption}
}
\description{
A function to plot densities of dead pixels of detector or module
}
