% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clump.R
\name{plot_events}
\alias{plot_events}
\title{Plots damaged layout events}
\usage{
plot_events(layout, file_path = NA, caption = TRUE,
  incl_event_list = NA, plot_edges_gaps = TRUE)
}
\arguments{
\item{layout}{Layout object}

\item{file_path}{Output file path}

\item{caption}{Flag to turn on/off figure caption}

\item{incl_event_list}{a list of events to be included}
}
\description{
Plots damaged layout events
}
