% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clump.R
\name{plot_events_angles}
\alias{plot_events_angles}
\title{Plots angles graph of events of a detector or module}
\usage{
plot_events_angles(
  detector,
  file_path = NA,
  row = NA,
  col = NA,
  caption = TRUE,
  incl_event_list = NA
)
}
\arguments{
\item{detector}{Detector object}

\item{file_path}{Output file path}

\item{row}{Module row number}

\item{col}{Module column number}

\item{caption}{Flag to turn on/off figure caption}

\item{incl_event_list}{a list of events to be included}
}
\description{
Plots angles graph of events of a detector or module
}
\examples{
detector_perkinfull <- create_detector("PerkinElmerFull")
file_path <-  system.file("extdata", "PerkinElmerFull",
                          "BadPixelMap_t1.bpm.xml",
                          package = "detectorchecker")
detector_perkinfull <- load_pix_matrix(
detector = detector_perkinfull, file_path = file_path)
detector_perkinfull_events = find_clumps(detector_perkinfull)
plot_events_angles(detector_perkinfull_events)
}
