% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clump.R
\name{plot_events_count}
\alias{plot_events_count}
\title{A function to plot layout with counts per module}
\usage{
plot_events_count(layout, file_path = NA, row = NA, col = NA,
  caption = TRUE, incl_event_list = NA)
}
\arguments{
\item{layout}{Layout object}

\item{file_path}{Output file path}

\item{row}{Module row number}

\item{col}{Module column number}

\item{caption}{Flag to turn on/off figure caption}
}
\description{
A function to plot layout with counts per module
}
