% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clump.R
\name{plot_events_kfg}
\alias{plot_events_kfg}
\title{Plots K, F, G functions of a detector or module}
\usage{
plot_events_kfg(detector, func, file_path = NA, row = NA, col = NA,
  caption = TRUE, incl_event_list = NA)
}
\arguments{
\item{detector}{Detector object}

\item{func}{Function name}

\item{file_path}{Output file path}

\item{row}{Module row number}

\item{col}{Module column number}

\item{caption}{Flag to turn on/off figure caption}

\item{incl_event_list}{a list of events to be included}
}
\description{
Plots K, F, G functions of a detector or module
}
