% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{plot_layout}
\alias{plot_layout}
\title{Plotting layout}
\usage{
plot_layout(layout, file_path = NA, caption = TRUE)
}
\arguments{
\item{layout}{Layout object}

\item{file_path}{Output file path}

\item{caption}{Flag to turn on/off figure caption}
}
\description{
Plotting layout
}
