% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{plot_layout_arrows}
\alias{plot_layout_arrows}
\title{A function to plot NN oriented arrrows}
\usage{
plot_layout_arrows(layout, file_path = NA, row = NA, col = NA,
  caption = TRUE)
}
\arguments{
\item{layout}{Layout object}

\item{file_path}{Output file path}

\item{row}{Module row number}

\item{col}{Module column number}

\item{caption}{Flag to turn on/off figure caption}
}
\description{
A function to plot NN oriented arrrows
}
