% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{plot_layout_damaged}
\alias{plot_layout_damaged}
\title{A function to plot layout with damaged pixels}
\usage{
plot_layout_damaged(layout, file_path = NA, caption = TRUE)
}
\arguments{
\item{layout}{Layout object}

\item{file_path}{Output file path}

\item{caption}{Flag to turn on/off figure caption}
}
\description{
A function to plot layout with damaged pixels
}
