% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{plot_layout_density}
\alias{plot_layout_density}
\title{A function to plot layout with dead pixel densities}
\usage{
plot_layout_density(layout, file_path = NA, adjust = 0.25, row = NA,
  col = NA, caption = TRUE)
}
\arguments{
\item{layout}{Layout object}

\item{file_path}{Output file path}

\item{adjust}{Kernel density bandwidth}

\item{row}{Module row number}

\item{col}{Module column number}

\item{caption}{Flag to turn on/off figure caption}
}
\description{
A function to plot layout with dead pixel densities
}
