% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{plot_layout_kfg}
\alias{plot_layout_kfg}
\title{Plots K, F, G functions}
\usage{
plot_layout_kfg(layout, func, file_path = NA, row = NA, col = NA,
  caption = TRUE)
}
\arguments{
\item{layout}{Layout object}

\item{func}{Function name}

\item{file_path}{Output file path}

\item{row}{module row number}

\item{col}{module column number}

\item{caption}{Flag to turn on/off figure caption}
}
\description{
Plots K, F, G functions
}
