% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{plot_layout_module_damaged}
\alias{plot_layout_module_damaged}
\title{Plotting a module of a layout}
\usage{
plot_layout_module_damaged(layout, col, row, file_path = NA,
  caption = TRUE)
}
\arguments{
\item{layout}{Layout object}

\item{col}{Module column number}

\item{row}{Module row number}

\item{file_path}{Output file path}

\item{caption}{Flag to turn on/off figure caption}
}
\description{
Plotting a module of a layout
}
