% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clump.R
\name{plot_module_events}
\alias{plot_module_events}
\title{Plots damaged detector module events}
\usage{
plot_module_events(detector, col, row, file_path = NA, caption = TRUE,
  incl_event_list = NA)
}
\arguments{
\item{detector}{Detector object}

\item{col}{Module column number}

\item{row}{Module row number}

\item{file_path}{Output file path}

\item{caption}{Flag to turn on/off figure caption}

\item{incl_event_list}{a list of events to be included}
}
\description{
Plots damaged detector module events
}
