% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pixel.R
\name{plot_pixel}
\alias{plot_pixel}
\title{Plots pixel analysis}
\usage{
plot_pixel(data, width, height, file_path = NA)
}
\arguments{
\item{data}{Matrix containing pixel analysis data}

\item{width}{Plot width}

\item{height}{Plot height}

\item{file_path}{Output path with an extension}
}
\description{
Plots pixel analysis
}
