% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pixel.R
\name{plot_pixel_dist_corner}
\alias{plot_pixel_dist_corner}
\title{Calculates and plots pixel distances from corners}
\usage{
plot_pixel_dist_corner(detector, file_path = NA)
}
\arguments{
\item{detector}{Detector object}

\item{file_path}{Output file path}
}
\description{
Calculates and plots pixel distances from corners
}
