% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pixel.R
\name{plot_pixel_dist_edge_row}
\alias{plot_pixel_dist_edge_row}
\title{Calculates and plots vetical distances from the module edges}
\usage{
plot_pixel_dist_edge_row(detector, file_path = NA)
}
\arguments{
\item{detector}{Detector object}

\item{file_path}{Output file path}
}
\description{
Calculates and plots vetical distances from the module edges
}
\examples{
# Create a detector
detector_pilatus <- create_detector("Pilatus")
# Load a pixel matrix
file_path <-  system.file("extdata", "Pilatus", "badpixel_mask.tif",
                         package ="detectorchecker")
detector_pilatus <- load_pix_matrix(detector = detector_pilatus, file_path = file_path)
plot_pixel_dist_edge_row(detector_pilatus)
}
