% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{plot_pixels_angles}
\alias{plot_pixels_angles}
\title{Plot nearest neighbour angles of dead pixels of detector or module}
\usage{
plot_pixels_angles(
  detector,
  file_path = NA,
  row = NA,
  col = NA,
  caption = TRUE
)
}
\arguments{
\item{detector}{Detector object}

\item{file_path}{Output file path}

\item{row}{Module row number}

\item{col}{Module column number}

\item{caption}{Flag to turn on/off figure caption}
}
\description{
Plot nearest neighbour angles of dead pixels of detector or module
}
\examples{
detector_perkinfull <- create_detector("PerkinElmerFull")
file_path <-  system.file("extdata", "PerkinElmerFull",
                          "BadPixelMap_t1.bpm.xml",
                          package = "detectorchecker")
detector_perkinfull <- load_pix_matrix(
detector = detector_perkinfull, file_path = file_path)
plot_pixels_angles(detector_perkinfull)
}
