% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{plot_pixels_count}
\alias{plot_pixels_count}
\title{A function to plot detector with dead pixel counts per module}
\usage{
plot_pixels_count(detector, file_path = NA, row = NA, col = NA, caption = TRUE)
}
\arguments{
\item{detector}{Detector object}

\item{file_path}{Output file path}

\item{row}{Module row number}

\item{col}{Module column number}

\item{caption}{Flag to turn on/off figure caption}
}
\description{
A function to plot detector with dead pixel counts per module
}
