% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{plot_pixels_density}
\alias{plot_pixels_density}
\title{A function to plot densities of dead pixels of detector or module}
\usage{
plot_pixels_density(
  detector,
  file_path = NA,
  adjust = 0.5,
  row = NA,
  col = NA,
  caption = TRUE,
  color = topo.colors(50)
)
}
\arguments{
\item{detector}{Detector object}

\item{file_path}{Output file path}

\item{adjust}{Kernel density bandwidth}

\item{row}{Module row number}

\item{col}{Module column number}

\item{caption}{Flag to turn on/off figure caption}

\item{color}{a list of colors}
}
\description{
A function to plot densities of dead pixels of detector or module
}
