% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clump.R
\name{plot_pixels_events}
\alias{plot_pixels_events}
\title{Plots damaged detector pixels and events}
\usage{
plot_pixels_events(detector, file_path = NA, caption = TRUE,
  incl_event_list = NA)
}
\arguments{
\item{detector}{Detector object}

\item{file_path}{Output file path}

\item{caption}{Flag to turn on/off figure caption}

\item{incl_event_list}{a list of events to be included}
}
\description{
Plots damaged detector pixels and events
}
