% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{plot_pixels_kfg}
\alias{plot_pixels_kfg}
\title{Plots K, F, G functions}
\usage{
plot_pixels_kfg(
  detector,
  func,
  file_path = NA,
  row = NA,
  col = NA,
  caption = TRUE
)
}
\arguments{
\item{detector}{Detector object}

\item{func}{Function name ("K', "F", or "G")}

\item{file_path}{Output file path}

\item{row}{module row number}

\item{col}{module column number}

\item{caption}{Flag to turn on/off figure caption}
}
\description{
Plots K, F, G functions
}
\examples{
# Create a detector
detector_pilatus <- create_detector("Pilatus")
# Load a pixel matrix
file_path <-  system.file("extdata", "Pilatus", "badpixel_mask.tif",
                         package ="detectorchecker")
detector_pilatus <- load_pix_matrix(detector = detector_pilatus, file_path = file_path)
plot_pixels_kfg(detector_pilatus, "K")
plot_pixels_kfg(detector_pilatus, "F")
plot_pixels_kfg(detector_pilatus, "G")
}
