% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{readin_detector}
\alias{readin_detector}
\title{Reads in a user defined detector from a file}
\usage{
readin_detector(file_path)
}
\arguments{
\item{file_path}{A path to the user defined detector file}
}
\value{
Detector object
}
\description{
Reads in a user defined detector from a file
}
