% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{readin_layout}
\alias{readin_layout}
\title{Reads in a user defined layout from a file}
\usage{
readin_layout(file_path)
}
\arguments{
\item{file_path}{A path to the user defined layout file}
}
\value{
Layout object
}
\description{
Reads in a user defined layout from a file
}
