% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{summary.detector}
\alias{summary.detector}
\title{Summary of a detector object}
\usage{
\method{summary}{detector}(detector)
}
\arguments{
\item{detector}{Detector object}
}
\value{
String with the detector summary
}
\description{
Summary of a detector object
}
\examples{
detc <- create_detector("Pilatus")
summary(detc)
}
