% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module.R
\name{which_module}
\alias{which_module}
\title{Module module
Returns row or column of a module that a dead pixel belongs to}
\usage{
which_module(coo, me)
}
\arguments{
\item{coo}{x or y coordinate of a dead pixel}

\item{me}{module edges}
}
\value{
row or column number
}
\description{
Module module
Returns row or column of a module that a dead pixel belongs to
}
