% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{check_detector_avail}
\alias{check_detector_avail}
\title{Checks whether \code{detector_name} is preconfigured.}
\usage{
check_detector_avail(detector_name)
}
\arguments{
\item{detector_name}{The name of the detector}
}
\value{
Boolean
}
\description{
If TRUE can be created by \code{create_detector}
}
\examples{
check_detector_avail('Pilatus')
}
