% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{create_detector}
\alias{create_detector}
\title{Create a Detector object}
\usage{
create_detector(detector_name)
}
\arguments{
\item{detector_name}{The name of the detector}
}
\value{
Detector S3 object
}
\description{
Create a Detector object.
If the \code{detector_name} is not available will raise an exception.
}
\examples{
detc <- create_detector(available_detectors[1])
}
