% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{create_module}
\alias{create_module}
\title{Checks whether layout is available, if so, creates a Layout object}
\usage{
create_module(layout_name)
}
\arguments{
\item{layout_name}{The name of the layout}
}
\value{
Layout object
}
\description{
Checks whether layout is available, if so, creates a Layout object
}
