% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{create_ppp_edges_col}
\alias{create_ppp_edges_col}
\title{TODO: Better description of the function
This is the ppp_edges_col function}
\usage{
create_ppp_edges_col(layout)
}
\arguments{
\item{layout}{Layout object}
}
\value{
Point pattern dataset
}
\description{
TODO: Better description of the function
This is the ppp_edges_col function
}
