% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{create_ppp_gaps_row}
\alias{create_ppp_gaps_row}
\title{This is the ppp_gaps_row creation function
This is a function}
\usage{
create_ppp_gaps_row(detector)
}
\arguments{
\item{detector}{Detector object}
}
\value{
Point pattern dataset
}
\description{
This is the ppp_gaps_row creation function
This is a function
}
