% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{dead_stats_summary}
\alias{dead_stats_summary}
\title{Summary of damaged pixels}
\usage{
dead_stats_summary(detector)
}
\arguments{
\item{detector}{Detector object}
}
\value{
A string with damaged pixels overall statitics
}
\description{
Summary of damaged pixels
}
