% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{Default_Layout}
\alias{Default_Layout}
\title{Layout module
A S3 class to represent a detector layout.}
\usage{
Default_Layout(name = "Default", date = NA, detector_width = NA,
  detector_height = NA, module_col_n = NA, module_row_n = NA,
  module_col_sizes = NA, module_row_sizes = NA, gap_col_sizes = NA,
  gap_row_sizes = NA, module_edges_col = NA, module_edges_row = NA,
  detector_inconsistency = NA, pix_matrix = NA, pix_dead = NA,
  dead_stats = NA, pix_dead_modules = NA, clumps = NA)
}
\arguments{
\item{name}{detector's name}

\item{date}{date}

\item{detector_width}{detector's width}

\item{detector_height}{detector's height}

\item{module_col_n}{number of columns in the grid of modules}

\item{module_row_n}{number of rows in the grid of modules}

\item{module_col_sizes}{vector with widths of the modules}

\item{module_row_sizes}{vector with heights of the modules}

\item{gap_col_sizes}{vector with widths of the gaps}

\item{gap_row_sizes}{vector with heights of the gaps}

\item{module_edges_col}{vector of columns that contain edges of modules}

\item{module_edges_row}{vector of rows that contain edges of modules}

\item{detector_inconsistency}{counts inconsistencies found in parameters entered}

\item{pix_matrix}{pixel matrix}

\item{pix_dead}{dead pixels coordinates}

\item{dead_stats}{dead pixel statistics}

\item{pix_dead_modules}{assigned module for each dead pixel}

\item{clumps}{clumps data (xyc_df data frame with pixels and their clump ID's, xyc_events data frame with clusters (clumps) and their clump ID's and centre coordinates)}
}
\value{
Layout object
}
\description{
Layout module
A S3 class to represent a detector layout.
}
