% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pixel.R
\name{dist_corner}
\alias{dist_corner}
\title{A function to calculate pixel distances from the closest corner}
\usage{
dist_corner(detector)
}
\arguments{
\item{detector}{Detector object}
}
\value{
Matrix containing pixel distances from closest corner
}
\description{
A function to calculate pixel distances from the closest corner
}
\examples{
detc <- Excalibur_exp_1
dist_corner(detc)
}
