% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pixel.R
\name{dist_corner}
\alias{dist_corner}
\title{A function to calculate pixel distances from the closest corner}
\usage{
dist_corner(detector)
}
\arguments{
\item{detector}{Detector object}
}
\value{
Matrix containing pixel distances from closest corner
}
\description{
A function to calculate pixel distances from the closest corner
}
\examples{
# Create a detector
detector_pilatus <- create_detector("Pilatus")
# Calculate distance from pixels to corners
dist_corner(detector_pilatus)
}
