% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pixel.R
\name{dist_edge_col}
\alias{dist_edge_col}
\title{A function to calculate pixel distances from edges by column}
\usage{
dist_edge_col(layout)
}
\arguments{
\item{layout}{Layout object}
}
\value{
dist ?
}
\description{
A function to calculate pixel distances from edges by column
}
