% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pixel.R
\name{dist_edge_col}
\alias{dist_edge_col}
\title{Calculate horizontal distance from each pixel to nearest module edge}
\usage{
dist_edge_col(detector)
}
\arguments{
\item{detector}{Detector object}
}
\value{
distance matrix
}
\description{
Calculate horizontal distance from each pixel to nearest module edge
}
\examples{
# Create a detector
detector_pilatus <- create_detector("Pilatus")
# Calculate horizontal distance from each pixel to nearest module edge
dist_edge_col(detector_pilatus)
}
