% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pixel.R
\name{dist_edge_min}
\alias{dist_edge_min}
\title{A function to calculate pixel distances from edges}
\usage{
dist_edge_min(detector)
}
\arguments{
\item{detector}{Detector object}
}
\value{
dist
}
\description{
A function to calculate pixel distances from edges
}
