% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pixel.R
\name{dist_edge_row}
\alias{dist_edge_row}
\title{A function to calculate pixel distances from edges by row}
\usage{
dist_edge_row(layout)
}
\arguments{
\item{layout}{Layout object}
}
\value{
dist ?
}
\description{
A function to calculate pixel distances from edges by row
}
