% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pixel.R
\name{dist_edge_row}
\alias{dist_edge_row}
\title{Calculate vertical distance from each pixel to nearest module edge}
\usage{
dist_edge_row(detector)
}
\arguments{
\item{detector}{Detector object}
}
\value{
distance matrix
}
\description{
Calculate vertical distance from each pixel to nearest module edge
}
\examples{
# Create a detector
detector_pilatus <- create_detector("Pilatus")
# Calculate vertical distance from each pixel to nearest module edge
dist_edge_row(detector_pilatus)
}
