% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clump.R
\name{.clump_module}
\alias{.clump_module}
\title{Identifying modules for clumps}
\usage{
.clump_module(detector, rrc)
}
\arguments{
\item{detector}{Detector object}

\item{rrc}{raster clumps objects}
}
\value{
data frame of the modules relating the clump
}
\description{
Identifying modules for clumps
}
