% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module.R
\name{.dist_edge}
\alias{.dist_edge}
\title{Function returns distance to both edges.
Given xy coo of pixel and matrices with positions of edges in both directions.}
\usage{
.dist_edge(xy, module_edges)
}
\arguments{
\item{xy}{}

\item{module_edges}{?}
}
\value{
tmp ?
}
\description{
Function returns distance to both edges.
Given xy coo of pixel and matrices with positions of edges in both directions.
}
