% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clump.R
\name{.get_clump_event_ppp}
\alias{.get_clump_event_ppp}
\title{Creates ppp object for damaged detector events}
\usage{
.get_clump_event_ppp(
  detector,
  incl_event_list = NA,
  height = NULL,
  width = NULL
)
}
\arguments{
\item{detector}{Detector object}

\item{incl_event_list}{a list of events to be included}

\item{height}{Detector height}

\item{width}{Detector width}
}
\value{
ppp object for damaged detector events
}
\description{
Creates ppp object for damaged detector events
}
\keyword{internal}
