% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clump.R
\name{.get_clump_event_ppp}
\alias{.get_clump_event_ppp}
\title{Creates ppp object for damaged detector events}
\usage{
.get_clump_event_ppp(
  detector,
  incl_event_list = NA,
  height = NULL,
  width = NULL
)
}
\arguments{
\item{detector}{Detector object}

\item{incl_event_list}{a list of events to be included}

\item{height}{Detector height}

\item{width}{Detector width}
}
\value{
ppp object for damaged detector events
}
\description{
Note that the parameter incl_event_list is a list of integers from 1-8.  For all events this would be list(1,2,3,4,5,6,7,8).
The integer values declare which damage events to include:
1=singletons, 2=doublets, 3=triplets, 4=larger clusters, 5=upper horizontal lines,
6=lower horizontal lines, 7=left vertical lines, 8=right vertical lines
}
\keyword{internal}
