% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clump.R
\name{.get_clump_pixel_ppp}
\alias{.get_clump_pixel_ppp}
\title{Creates ppp for damaged detector pixels}
\usage{
.get_clump_pixel_ppp(detector, incl_event_list = NA)
}
\arguments{
\item{detector}{Detector object}

\item{incl_event_list}{a list of events to be included}
}
\description{
Creates ppp for damaged detector pixels
}
