% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clump.R
\name{.xyc_pixels2events}
\alias{.xyc_pixels2events}
\title{MODIFYING CLUSTERS TO EVENTS (CONSISTING OF 1 PIXEL REPRESENTING THE CLUSTER)
Make into a point pattern of just events rather than pixels. Using xyc_ply object.
Collapse in one point using centres for clusters, but end points for lines, type dependend:
type 5 (closest to upper edge): ymin
type 6 (closest to lower edge): ymax
type 7 (closest to right edge): xmin
type 8 (closest to left edge):  xmax
This is inspired by Perkin Elmer Detector and be replaced by other choices if desired.}
\usage{
.xyc_pixels2events(xyc_ply)
}
\arguments{
\item{xyc_ply}{clums data frame}
}
\value{
events
}
\description{
MODIFYING CLUSTERS TO EVENTS (CONSISTING OF 1 PIXEL REPRESENTING THE CLUSTER)
Make into a point pattern of just events rather than pixels. Using xyc_ply object.
Collapse in one point using centres for clusters, but end points for lines, type dependend:
type 5 (closest to upper edge): ymin
type 6 (closest to lower edge): ymax
type 7 (closest to right edge): xmin
type 8 (closest to left edge):  xmax
This is inspired by Perkin Elmer Detector and be replaced by other choices if desired.
}
