% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clump.R
\name{.xyc_ply_func}
\alias{.xyc_ply_func}
\title{Something something dark side}
\usage{
.xyc_ply_func(layout, xyc_pixel_df)
}
\arguments{
\item{layout}{Layout object}
}
\value{
data frame
}
\description{
Something something dark side
}
