% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clump.R
\name{find_clumps}
\alias{find_clumps}
\title{Locates and clasifies clumps of a damaged detector}
\usage{
find_clumps(detector, row = NA, col = NA)
}
\arguments{
\item{detector}{Detector object}

\item{row}{Module row number}

\item{col}{Module column number}
}
\value{
Detector with events matrix
}
\description{
Locates and clasifies clumps of a damaged detector
}
