% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clump.R
\name{find_clumps}
\alias{find_clumps}
\title{Locates and clusifies clumps of a damaged layout}
\usage{
find_clumps(layout, row = NA, col = NA)
}
\arguments{
\item{layout}{Layout object}

\item{row}{Module row number}

\item{col}{Module column number}
}
\description{
Locates and clusifies clumps of a damaged layout
}
