% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pixel.R
\name{get_dead_pix_mask}
\alias{get_dead_pix_mask}
\title{Creates a mask matrix of dead pixels}
\usage{
get_dead_pix_mask(detector)
}
\arguments{
\item{detector}{Detector object}
}
\value{
dead pixel mask
}
\description{
Converts the pix_dead attribute of a detector (NX2 list) to a matrix of 1 and 0 (1 for dead pixel)
}
\examples{
detc <- Excalibur_exp_1
}
