% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pixel.R
\name{get_dead_pix_mask}
\alias{get_dead_pix_mask}
\title{Creates a mask matrix of dead pixels}
\usage{
get_dead_pix_mask(layout)
}
\arguments{
\item{layout}{Layout object}
}
\value{
dead pixel mask
}
\description{
Creates a mask matrix of dead pixels
}
