% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{get_dead_stats}
\alias{get_dead_stats}
\title{Generate summary of damaged pixels and add as a dead_stats attribute to the detector object}
\usage{
get_dead_stats(detector)
}
\arguments{
\item{detector}{Detector object}
}
\value{
Detector object with dead_stats attribute
}
\description{
Generate summary of damaged pixels and add as a dead_stats attribute to the detector object
}
\examples{
detc <- Excalibur_exp_1
get_dead_stats(detc)$dead_stats
}
