% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{get_dead_stats}
\alias{get_dead_stats}
\title{Count number of damaged pixels overall and in different modules}
\usage{
get_dead_stats(layout)
}
\arguments{
\item{layout}{Layout object}
}
\value{
Dead_Stats object
}
\description{
Count number of damaged pixels overall and in different modules
}
