% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clump.R
\name{get_events_mask}
\alias{get_events_mask}
\title{Generates events matrix (a matrix with pixels as 0 and events as 1)}
\usage{
get_events_mask(layout)
}
\arguments{
\item{layout}{Layout object}
}
\value{
events mask
}
\description{
Generates events matrix (a matrix with pixels as 0 and events as 1)
}
