% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{get_ppp_dead}
\alias{get_ppp_dead}
\title{Generates ppp for the dead pixels}
\usage{
get_ppp_dead(detector)
}
\arguments{
\item{detector}{Detector object}
}
\value{
ppp of dead pixels
}
\description{
Generates ppp for the dead pixels
}
