% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clump.R
\name{glm_events_dist_corner}
\alias{glm_events_dist_corner}
\title{Fits events distances to the nearest corner using glm}
\usage{
glm_events_dist_corner(detector, incl_event_list = NA)
}
\arguments{
\item{detector}{Detector object}

\item{incl_event_list}{a list of events to be included}
}
\value{
Fitted model
}
\description{
Note that the parameter incl_event_list is a list of integers from \link{1,8}.  For all events this would be list(1,2,3,4,5,6,7,8).
The integer values declare which damage events to include:
1=singletons, 2=doublets, 3=triplets, 4=larger clusters, 5=upper horizontal lines,
6=lower horizontal lines, 7=left vertical lines, 8=right vertical lines
}
