% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{glm_pixel_ctr_eucl}
\alias{glm_pixel_ctr_eucl}
\title{Fits pixel distance from the centre using glm}
\usage{
glm_pixel_ctr_eucl(detector)
}
\arguments{
\item{detector}{Detector object}
}
\value{
Fitted model
}
\description{
Fits pixel distance from the centre using glm
}
