% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{glm_pixel_ctr_linf}
\alias{glm_pixel_ctr_linf}
\title{Predict dead pixels from the pixel's parallel maxima}
\usage{
glm_pixel_ctr_linf(detector)
}
\arguments{
\item{detector}{Detector object}
}
\value{
Fitted model
}
\description{
Fit a logistic regression model using \code{glm}.
Predicts dead pixels from the pixel's parallel maxima
}
\examples{
glm_pixel_ctr_linf(PerkinElmerFull_exp_1)
}
