% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{glm_pixel_ctr_linf}
\alias{glm_pixel_ctr_linf}
\title{Fits pixel parallel maxima from the centre}
\usage{
glm_pixel_ctr_linf(detector)
}
\arguments{
\item{detector}{Detector object}
}
\value{
Fitted model
}
\description{
Fits pixel parallel maxima from the centre
}
