% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{glm_pixel_dist_corner}
\alias{glm_pixel_dist_corner}
\title{Fits pixel distances to the nearest corner using glm}
\usage{
glm_pixel_dist_corner(detector)
}
\arguments{
\item{detector}{Detector object}
}
\value{
Fitted model
}
\description{
Fits pixel distances to the nearest corner using glm
}
