% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{glm_pixel_dist_edge_col}
\alias{glm_pixel_dist_edge_col}
\title{Fits pixel distances from the module edges by column using glm}
\usage{
glm_pixel_dist_edge_col(detector)
}
\arguments{
\item{detector}{Detector object}
}
\value{
Fitted model
}
\description{
Fits pixel distances from the module edges by column using glm
}
