% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{glm_pixel_dist_edge_row}
\alias{glm_pixel_dist_edge_row}
\title{Predict dead pixels from pixel distances from the module edges by module row}
\usage{
glm_pixel_dist_edge_row(detector)
}
\arguments{
\item{detector}{Detector object}
}
\value{
Fitted model
}
\description{
Fit a logistic regression model using \code{glm}.
Predict dead pixels from pixel distances from the module edges by module row
}
\examples{
glm_pixel_dist_edge_row(PerkinElmerFull_exp_1)
}
