% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{load_pix_matrix}
\alias{load_pix_matrix}
\title{A function to load pixel data}
\usage{
load_pix_matrix(layout, file_path)
}
\arguments{
\item{layout}{The name of the layout to be used}

\item{file_path}{Path(s) to the file(s) containing dead pixel information}
}
\value{
Layout object
}
\description{
A function to load pixel data
}
