% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{load_pix_matrix}
\alias{load_pix_matrix}
\title{A function to load pixel data}
\usage{
load_pix_matrix(detector, file_path)
}
\arguments{
\item{detector}{The name of the detector object to be used}

\item{file_path}{Path(s) to the file(s) containing dead pixel information}
}
\value{
Detector object
}
\description{
A function to load pixel data
}
