% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{PerkinElmerCropped1600_Layout}
\alias{PerkinElmerCropped1600_Layout}
\title{A S3 class to represent the PerkinElmerCropped1600 detector layout.}
\usage{
PerkinElmerCropped1600_Layout()
}
\value{
PerkinElmerCropped1600 layout object
}
\description{
A S3 class to represent the PerkinElmerCropped1600 detector layout.
}
