% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pixel.R
\name{pixel_dist_ctr_eucl}
\alias{pixel_dist_ctr_eucl}
\title{A function to calculate euclidean distance from the centre for each pixel}
\usage{
pixel_dist_ctr_eucl(detector)
}
\arguments{
\item{detector}{Detector object}
}
\value{
Matrix of euclidean distances
}
\description{
A function to calculate euclidean distance from the centre for each pixel
}
