% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pixel.R
\name{pixel_dist_ctr_eucl}
\alias{pixel_dist_ctr_eucl}
\title{Calculate euclidean distance from the centre of a module for each pixel}
\usage{
pixel_dist_ctr_eucl(detector)
}
\arguments{
\item{detector}{Detector object}
}
\value{
Matrix of euclidean distances
}
\description{
Calculate euclidean distance from the centre of a module for each pixel
}
\examples{
pixel_dist_ctr_eucl(PerkinElmerFull_Detector())
}
