% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{plot_detector_angles}
\alias{plot_detector_angles}
\title{A function to plot NN angles of dead pixels of detector or module}
\usage{
plot_detector_angles(
  detector,
  file_path = NA,
  row = NA,
  col = NA,
  caption = TRUE
)
}
\arguments{
\item{detector}{Detector object}

\item{file_path}{Output file path}

\item{row}{Module row number}

\item{col}{Module column number}

\item{caption}{Flag to turn on/off figure caption}
}
\description{
A function to plot NN angles of dead pixels of detector or module
}
