% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clump.R
\name{plot_events}
\alias{plot_events}
\title{Plots damaged detector events}
\usage{
plot_events(
  detector,
  file_path = NA,
  caption = TRUE,
  incl_event_list = NA,
  plot_edges_gaps = TRUE
)
}
\arguments{
\item{detector}{Detector object}

\item{file_path}{Output file path}

\item{caption}{Flag to turn on/off figure caption}

\item{incl_event_list}{a list of events to be included}

\item{plot_edges_gaps}{Plot edgees gaps}
}
\description{
Plots damaged detector events
}
