% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clump.R
\name{plot_events_angles}
\alias{plot_events_angles}
\title{Plots angles graph of events of a detector or module}
\usage{
plot_events_angles(
  detector,
  file_path = NA,
  row = NA,
  col = NA,
  caption = TRUE,
  incl_event_list = NA
)
}
\arguments{
\item{detector}{Detector object}

\item{file_path}{Output file path}

\item{row}{Module row number}

\item{col}{Module column number}

\item{caption}{Flag to turn on/off figure caption}

\item{incl_event_list}{a list of events to be included}
}
\description{
Note that the parameter incl_event_list is a list of integers from [1,8].  For all events this would be list(1,2,3,4,5,6,7,8).
The integer values declare which damage events to include:
     1=singletons, 2=doublets, 3=triplets, 4=larger clusters, 5=upper horizontal lines,
     6=lower horizontal lines, 7=left vertical lines, 8=right vertical lines
}
\examples{
detector_perkinfull <- create_detector("PerkinElmerFull")
file_path <-  system.file("extdata", "PerkinElmerFull",
                          "BadPixelMap_t1.bpm.xml",
                          package = "detectorchecker")
detector_perkinfull <- load_pix_matrix(
detector = detector_perkinfull, file_path = file_path)
detector_perkinfull_events = find_clumps(detector_perkinfull)
plot_events_angles(detector_perkinfull_events)
}
