% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clump.R
\name{plot_events_density}
\alias{plot_events_density}
\title{Plots density graph of events}
\usage{
plot_events_density(layout, file_path = NA, adjust = 0.25, row = NA,
  col = NA, caption = TRUE, incl_event_list = NA)
}
\arguments{
\item{layout}{Layout object}

\item{file_path}{Output file path}

\item{adjust}{Kernel density bandwidth}

\item{row}{Module row number}

\item{col}{Module column number}

\item{caption}{Flag to turn on/off figure caption}

\item{incl_event_list}{a list of events to be included}
}
\description{
Plots density graph of events
}
