% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clump.R
\name{plot_module_events}
\alias{plot_module_events}
\title{Plots damaged detector module events}
\usage{
plot_module_events(
  detector,
  col,
  row,
  file_path = NA,
  caption = TRUE,
  incl_event_list = NA
)
}
\arguments{
\item{detector}{Detector object}

\item{col}{Module column number}

\item{row}{Module row number}

\item{file_path}{Output file path}

\item{caption}{Flag to turn on/off figure caption}

\item{incl_event_list}{a list of events to be included}
}
\description{
Plots damaged detector module events
}
\examples{
# Create a detector
detector_pilatus <- create_detector("Pilatus")
# Load a pixel matrix
file_path <-  system.file("extdata", "Pilatus", "badpixel_mask.tif",
                         package ="detectorchecker")
detector_pilatus <- load_pix_matrix(detector = detector_pilatus, file_path = file_path)
 #Plot module events
 plot_module_events(detector_pilatus, 1, 1)
}
