% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clump.R
\name{plot_module_events}
\alias{plot_module_events}
\title{Plots damaged detector module events}
\usage{
plot_module_events(
  detector,
  col,
  row,
  file_path = NA,
  caption = TRUE,
  incl_event_list = NA
)
}
\arguments{
\item{detector}{Detector object}

\item{col}{Module column number}

\item{row}{Module row number}

\item{file_path}{Output file path}

\item{caption}{Flag to turn on/off figure caption}

\item{incl_event_list}{a list of events to be included}
}
\description{
Note that the parameter incl_event_list is a list of integers from \link{1,8}.  For all events this would be list(1,2,3,4,5,6,7,8).
The integer values declare which damage events to include:
1=singletons, 2=doublets, 3=triplets, 4=larger clusters, 5=upper horizontal lines,
6=lower horizontal lines, 7=left vertical lines, 8=right vertical lines
}
\examples{
# Create a detector
detector_pilatus <- create_detector("Pilatus")
# Load a pixel matrix
file_path <-  system.file("extdata", "Pilatus", "badpixel_mask.tif",
                         package ="detectorchecker")
detector_pilatus <- load_pix_matrix(detector = detector_pilatus, file_path = file_path)
 #Plot module events
 plot_module_events(detector_pilatus, 1, 1)
}
