% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{plot_module_pixels}
\alias{plot_module_pixels}
\title{A function to plot detector module with damaged pixels}
\usage{
plot_module_pixels(detector, col, row, file_path = NA, caption = TRUE)
}
\arguments{
\item{detector}{Detector object}

\item{col}{Module column number}

\item{row}{Module row number}

\item{file_path}{Output file path}

\item{caption}{Flag to turn on/off figure caption}
}
\description{
A function to plot detector module with damaged pixels
}
