% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pixel.R
\name{plot_pixel_ctr_eucl}
\alias{plot_pixel_ctr_eucl}
\title{Calculates and plots pixel euclidean distance from the centre}
\usage{
plot_pixel_ctr_eucl(detector, file_path = NA)
}
\arguments{
\item{detector}{Detector object}

\item{file_path}{Output file path}
}
\description{
Calculates and plots pixel euclidean distance from the centre
}
