% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pixel.R
\name{plot_pixel_ctr_linf}
\alias{plot_pixel_ctr_linf}
\title{Calculates and plots pixel parallel maxima from the centre}
\usage{
plot_pixel_ctr_linf(layout, file_path = NA)
}
\arguments{
\item{layout}{Layout object}

\item{file_path}{Output file path}
}
\description{
Calculates and plots pixel parallel maxima from the centre
}
