% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pixel.R
\name{plot_pixel_dist_edge}
\alias{plot_pixel_dist_edge}
\title{Calculates and plots minimum distances from the module edges}
\usage{
plot_pixel_dist_edge(layout, file_path = NA)
}
\arguments{
\item{layout}{Layout object}

\item{file_path}{Output file path}
}
\description{
Calculates and plots minimum distances from the module edges
}
