% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pixel.R
\name{plot_pixel_dist_edge_row}
\alias{plot_pixel_dist_edge_row}
\title{Calculates and plots distances from the module edges by row}
\usage{
plot_pixel_dist_edge_row(detector, file_path = NA)
}
\arguments{
\item{detector}{Detector object}

\item{file_path}{Output file path}
}
\description{
Calculates and plots distances from the module edges by row
}
