% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{plot_pixels}
\alias{plot_pixels}
\title{A function to plot detector with damaged pixels overlayed}
\usage{
plot_pixels(detector, col = NA, row = NA, file_path = NA, caption = TRUE)
}
\arguments{
\item{detector}{Detector object}

\item{col}{Module column number}

\item{row}{Module row number}

\item{file_path}{Output file path}

\item{caption}{Flag to turn on/off figure caption}
}
\description{
A function to plot detector with damaged pixels overlayed
}
\examples{
# Create a detector
detector_pilatus <- create_detector("Pilatus")
# Load a pixel matrix
file_path <-  system.file("extdata", "Pilatus", "badpixel_mask.tif",
                         package ="detectorchecker")
detector_pilatus <- load_pix_matrix(detector = detector_pilatus, file_path = file_path)
plot_pixels(detector_pilatus)
plot_pixels(detector_pilatus, row = 1, col = 1)
}
