% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pixel.R
\name{remove_high_density_cluster}
\alias{remove_high_density_cluster}
\title{Remove high density cluster of dead pixels
Recalculates dead statistics and clumps if they were present in the Detector object}
\usage{
remove_high_density_cluster(detector, min_pts = 30, eps_adjust = 0.05)
}
\arguments{
\item{detector}{Detector object}

\item{min_pts}{minimum points argument of dbscan function}

\item{eps_adjust}{}
}
\value{
detector object with high density cluster of pixels removed
}
\description{
Remove high density cluster of dead pixels
Recalculates dead statistics and clumps if they were present in the Detector object
}
