% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module.R
\name{which_module}
\alias{which_module}
\title{Module module
Which module function}
\usage{
which_module(coo, me)
}
\arguments{
\item{coo}{?}

\item{me}{?}
}
\value{
which_module what does this mean?
}
\description{
Module module
Which module function
}
