% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module.R
\name{which_module_idx}
\alias{which_module_idx}
\title{Function returns both col and row wrt detector grid.
Given xy coo of pixel and matrices with positions of edges in both directions.}
\usage{
which_module_idx(x, y, module_edges_col, module_edges_row)
}
\arguments{
\item{x}{}

\item{y}{?}

\item{module_edges_col}{?}

\item{module_edges_row}{?}
}
\value{
tmp ?
}
\description{
Function returns both col and row wrt detector grid.
Given xy coo of pixel and matrices with positions of edges in both directions.
}
