% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module.R
\name{which_module_idx}
\alias{which_module_idx}
\title{Function returns both col and row number of a dead pixel.}
\usage{
which_module_idx(x, y, module_edges_col, module_edges_row)
}
\arguments{
\item{x}{pixel x coordinate}

\item{y}{pixel y coordinate}

\item{module_edges_col}{vector of columns that contain edges of modules}

\item{module_edges_row}{vector of rows that contain edges of modules}
}
\value{
tmp
}
\description{
Function returns both col and row number of a dead pixel.
}
